refactor: clean up routes
